Move local message controls under local agent
