+++ packages/cli/src/cli/commands/local-agent.test.ts
@@ -16,6 +16,8 @@ function harness(overrides: Partial<LocalAgentDependencies> = {}) {
     listAgents: vi.fn(async () => [{ name: 'lead' }]),
     release: vi.fn(async () => undefined),
     setModel: vi.fn(async () => ({ name: 'lead', model: 'opus', success: true })),
+    flushPending: vi.fn(async () => ({ flushed: 2 })),
+    setInboundDeliveryMode: vi.fn(async (_name: string, mode: string) => ({ mode, flushed: 0 })),
   };
   const attach = vi.fn(async () => 0);
   const log = vi.fn();
@@ -97,5 +99,66 @@ describe('local agent subtree', () => {
     const { program, client } = harness();
     await program.parseAsync(['local', 'agent', 'set-model', 'lead', 'opus'], { from: 'user' });
     expect(client.setModel).toHaveBeenCalledWith('lead', 'opus');
+  });
+
+  it('message flush drains a local broker agent queue', async () => {
+    const client = { flushPending: vi.fn(async () => ({ flushed: 2 })) };
+    const connectLocal = vi.fn(async () => client as never);
+    const { program, log } = harness({ connectLocal });
+
+    await program.parseAsync(
+      [
+        'local',
+        'agent',
+        'message',
+        'flush',
+        'claude',
+        '--broker-url',
+        'http://127.0.0.1:3890',
+        '--api-key',
+        'secret',
+        '--state-dir',
+        '/tmp/relay-state',
+      ],
+      { from: 'user' }
+    );
+
+    expect(connectLocal).toHaveBeenCalledWith('/tmp/project', {
+      brokerUrl: 'http://127.0.0.1:3890',
+      apiKey: 'secret',
+      stateDir: '/tmp/relay-state',
+    });
+    expect(client.flushPending).toHaveBeenCalledWith('claude');
+    expect(log).toHaveBeenCalledWith(JSON.stringify({ name: 'claude', flushed: 2 }, null, 2));
+  });
+
+  it('message hold and auto switch local broker delivery mode', async () => {
+    const client = {
+      setInboundDeliveryMode: vi.fn(async (_name: string, mode: string) => ({ mode, flushed: 0 })),
+    };
+    const connectLocal = vi.fn(async () => client as never);
+    const { program, log } = harness({ connectLocal });
+
+    await program.parseAsync(['local', 'agent', 'message', 'hold', 'claude'], { from: 'user' });
+    await program.parseAsync(['local', 'agent', 'message', 'auto', 'claude'], { from: 'user' });
+
+    expect(connectLocal).toHaveBeenNthCalledWith(1, '/tmp/project', {
+      brokerUrl: undefined,
+      apiKey: undefined,
+      stateDir: undefined,
+    });
+    expect(connectLocal).toHaveBeenNthCalledWith(2, '/tmp/project', {
+      brokerUrl: undefined,
+      apiKey: undefined,
+      stateDir: undefined,
+    });
+    expect(client.setInboundDeliveryMode).toHaveBeenNthCalledWith(1, 'claude', 'manual_flush');
+    expect(client.setInboundDeliveryMode).toHaveBeenNthCalledWith(2, 'claude', 'auto_inject');
+    expect(log).toHaveBeenCalledWith(
+      JSON.stringify({ name: 'claude', mode: 'manual_flush', flushed: 0 }, null, 2)
+    );
+    expect(log).toHaveBeenCalledWith(
+      JSON.stringify({ name: 'claude', mode: 'auto_inject', flushed: 0 }, null, 2)
+    );
   });
 });
