+++ packages/cli/src/cli/commands/relaycast-groups.test.ts
@@ -1,7 +1,7 @@
 import { Command } from 'commander';
 import { describe, expect, it, vi } from 'vitest';
 
-import { registerAgentCommands, type AgentCommandDependencies } from './agent.js';
+import { registerAgentCommands } from './agent.js';
 import { registerChannelCommands } from './channel.js';
 import { registerMessageCommands } from './message.js';
 import { registerIntegrationCommands } from './integration.js';
@@ -74,138 +74,6 @@ describe('SDK-backed CLI groups', () => {
       expect.objectContaining({ name: 'reviewer', type: 'agent' })
     );
     expect(log).toHaveBeenCalled();
-  });
-
-  it('agent message flush drains a local broker agent queue', async () => {
-    const relay = createRelayMock();
-    const client = { flushPending: vi.fn(async () => ({ flushed: 2 })) };
-    const log = vi.fn();
-    const deps: Partial<AgentCommandDependencies> = {
-      createAgentRelay: () => relay as never,
-      createWorkspaceRelay: () => relay as never,
-      connectLocal: vi.fn(async () => client as never),
-      cwd: () => '/tmp/project',
-      log,
-      error: vi.fn(),
-      exit: vi.fn() as never,
-    };
-    const program = new Command();
-    program.exitOverride();
-    registerAgentCommands(program, deps);
-
-    await program.parseAsync(
-      [
-        'agent',
-        'message',
-        'flush',
-        'claude',
-        '--broker-url',
-        'http://127.0.0.1:3890',
-        '--api-key',
-        'secret',
-        '--state-dir',
-        '/tmp/relay-state',
-      ],
-      { from: 'user' }
-    );
-
-    expect(deps.connectLocal).toHaveBeenCalledWith('/tmp/project', {
-      brokerUrl: 'http://127.0.0.1:3890',
-      apiKey: 'secret',
-      stateDir: '/tmp/relay-state',
-    });
-    expect(client.flushPending).toHaveBeenCalledWith('claude');
-    expect(log).toHaveBeenCalledWith(JSON.stringify({ name: 'claude', flushed: 2 }, null, 2));
-  });
-
-  it('agent message controls resolve broker selection flags with the default local client', async () => {
-    const relay = createRelayMock();
-    const fetch = vi.fn(async () => new Response(JSON.stringify({ flushed: 3 }), { status: 200 }));
-    const readConnectionFile = vi.fn(() => ({ url: 'http://file-broker:1111', api_key: 'file-key' }));
-    const log = vi.fn();
-    const deps: Partial<AgentCommandDependencies> = {
-      createAgentRelay: () => relay as never,
-      createWorkspaceRelay: () => relay as never,
-      cwd: () => '/tmp/project',
-      readConnectionFile,
-      getDefaultStateDir: () => '/tmp/default-state',
-      env: {},
-      fetch: fetch as never,
-      log,
-      error: vi.fn(),
-      exit: vi.fn() as never,
-    };
-    const program = new Command();
-    program.exitOverride();
-    registerAgentCommands(program, deps);
-
-    await program.parseAsync(
-      [
-        'agent',
-        'message',
-        'flush',
-        'claude',
-        '--broker-url',
-        'http://flag-broker:2222/',
-        '--api-key',
-        'flag-key',
-        '--state-dir',
-        '/tmp/relay-state',
-      ],
-      { from: 'user' }
-    );
-
-    expect(readConnectionFile).toHaveBeenCalledWith('/tmp/relay-state');
-    expect(fetch).toHaveBeenCalledWith(
-      'http://flag-broker:2222/api/spawned/claude/flush',
-      expect.objectContaining({
-        method: 'POST',
-        headers: expect.objectContaining({ 'X-API-Key': 'flag-key' }),
-      })
-    );
-    expect(log).toHaveBeenCalledWith(JSON.stringify({ name: 'claude', flushed: 3 }, null, 2));
-  });
-
-  it('agent message hold and auto switch local broker delivery mode', async () => {
-    const relay = createRelayMock();
-    const client = {
-      setInboundDeliveryMode: vi.fn(async (_name: string, mode: string) => ({ mode, flushed: 0 })),
-    };
-    const log = vi.fn();
-    const deps: Partial<AgentCommandDependencies> = {
-      createAgentRelay: () => relay as never,
-      createWorkspaceRelay: () => relay as never,
-      connectLocal: vi.fn(async () => client as never),
-      cwd: () => '/tmp/project',
-      log,
-      error: vi.fn(),
-      exit: vi.fn() as never,
-    };
-    const program = new Command();
-    program.exitOverride();
-    registerAgentCommands(program, deps);
-
-    await program.parseAsync(['agent', 'message', 'hold', 'claude'], { from: 'user' });
-    await program.parseAsync(['agent', 'message', 'auto', 'claude'], { from: 'user' });
-
-    expect(deps.connectLocal).toHaveBeenNthCalledWith(1, '/tmp/project', {
-      brokerUrl: undefined,
-      apiKey: undefined,
-      stateDir: undefined,
-    });
-    expect(deps.connectLocal).toHaveBeenNthCalledWith(2, '/tmp/project', {
-      brokerUrl: undefined,
-      apiKey: undefined,
-      stateDir: undefined,
-    });
-    expect(client.setInboundDeliveryMode).toHaveBeenNthCalledWith(1, 'claude', 'manual_flush');
-    expect(client.setInboundDeliveryMode).toHaveBeenNthCalledWith(2, 'claude', 'auto_inject');
-    expect(log).toHaveBeenCalledWith(
-      JSON.stringify({ name: 'claude', mode: 'manual_flush', flushed: 0 }, null, 2)
-    );
-    expect(log).toHaveBeenCalledWith(
-      JSON.stringify({ name: 'claude', mode: 'auto_inject', flushed: 0 }, null, 2)
-    );
   });
 
   it('channel set_topic calls channels.update', async () => {
